+++ src/background/cli/self_pipe.rs
@@ -4,11 +4,11 @@ use slu_ipc::{
     AppIpc,
 };
 
-use crate::{cli::application::AppCli, error::Result};
+use crate::{cli::application::AppCli, error::Result, modules::system_tray::SystemTrayManager};
 
 pub struct SelfPipe;
 impl SelfPipe {
-    fn _handle_message(mut argv: Vec<String>) -> Result<()> {
+    fn _handle_cli_message(mut argv: Vec<String>) -> Result<()> {
         if argv.is_empty() {
             return Ok(());
         }
@@ -27,10 +27,20 @@ impl SelfPipe {
         Ok(())
     }
 
-    fn handle_message(argv: Vec<String>) -> IpcResponse {
-        match Self::_handle_message(argv) {
-            Ok(()) => IpcResponse::Success,
-            Err(err) => IpcResponse::Err(err.to_string()),
+    fn handle_message(message: AppMessage) -> IpcResponse {
+        match message {
+            AppMessage::Cli(argv) => match Self::_handle_cli_message(argv) {
+                Ok(()) => IpcResponse::Success,
+                Err(err) => IpcResponse::Err(err.to_string()),
+            },
+            AppMessage::TrayChanged(event) => {
+                SystemTrayManager::handle_tray_event(event);
+                IpcResponse::Success
+            }
+            AppMessage::Debug(_msg) => {
+                // log::debug!("{_msg}");
+                IpcResponse::Success
+            }
         }
     }
 
@@ -40,7 +50,7 @@ impl SelfPipe {
     }
 
     pub async fn request_open_settings() -> Result<()> {
-        AppIpc::send(AppMessage(vec!["settings".to_owned()])).await?;
+        AppIpc::send(AppMessage::Cli(vec!["settings".to_owned()])).await?;
         Ok(())
     }
 }