+++ src/background/modules/system_tray/application/mod.rs
@@ -1,5 +1,5 @@
+pub mod tray_hook_loader;
 pub mod tray_icon;
-pub mod tray_spy;
 mod util;
 
 use std::sync::LazyLock;
@@ -8,49 +8,53 @@ use seelen_core::{
     handlers::SeelenEvent,
     system_state::{SysTrayIcon, SysTrayIconId},
 };
+use slu_ipc::messages::Win32TrayEvent;
 use tauri::Emitter;
 
 use crate::{
     app::get_app_handle,
     error::{ErrorMap, ResultLogExt},
-    modules::system_tray::application::tray_spy::TraySpy,
-    utils::{lock_free::SyncHashMap, spawn_named_thread},
+    modules::system_tray::application::tray_hook_loader::TrayHookLoader,
+    utils::lock_free::SyncHashMap,
 };
 
 static SYSTEM_TRAY_MANAGER: LazyLock<SystemTrayManager> = LazyLock::new(SystemTrayManager::create);
 
 pub struct SystemTrayManager {
     icons: SyncHashMap<SysTrayIconId, SysTrayIcon>,
-    _spy: Option<TraySpy>,
+    _loader: Option<TrayHookLoader>,
 }
 
 impl SystemTrayManager {
     fn create() -> Self {
-        let Ok((_spy, mut event_rx)) = TraySpy::new() else {
-            return Self {
-                icons: SyncHashMap::new(),
-                _spy: None,
-            };
-        };
+        log::trace!("Creating system tray manager");
 
-        spawn_named_thread("System Tray", move || {
-            while let Some(event) = event_rx.blocking_recv() {
-                if let Some(_event) = Self::instance().process_event(event) {
-                    get_app_handle()
-                        .emit(SeelenEvent::SystemTrayChanged, Self::instance().icons())
-                        .wrap_error()
-                        .log_error();
-                }
+        let loader = match TrayHookLoader::new() {
+            Ok(loader) => Some(loader),
+            Err(err) => {
+                log::error!("Failed to create tray hook loader: {:?}", err);
+                None
             }
-        });
+        };
 
         Self {
             icons: SyncHashMap::new(),
-            _spy: Some(_spy),
+            _loader: loader,
         }
     }
 
     pub fn instance() -> &'static Self {
         &SYSTEM_TRAY_MANAGER
+    }
+
+    /// Handles a tray event received via IPC
+    /// This method should be called from the AppIpc handler
+    pub fn handle_tray_event(event: Win32TrayEvent) {
+        if let Some(_event) = Self::instance().process_event(event) {
+            get_app_handle()
+                .emit(SeelenEvent::SystemTrayChanged, Self::instance().icons())
+                .wrap_error()
+                .log_error();
+        }
     }
 }