+++ src/background/modules/system_tray/mod.rs
@@ -1,3 +1,5 @@
 mod application;
 mod domain;
 pub mod infrastructure;
+
+pub use application::SystemTrayManager;