+++ src/background/modules/tray (old)/mod.rs
@@ -1,7 +1,0 @@
-//! this code is here just for read-only purpose, tray module is no longer used.
-//! we decided to stop supporting tray items because of its instability and imposibility of
-//! be correctly implemented.
-
-pub mod application;
-pub mod domain;
-pub mod infrastructure;